Merge feature/11_TeamSession

+++ src/main/java/com/teamstation/teamstation/controller/ProjectTodoController.java
@@ -65,6 +65,12 @@ public class ProjectTodoController {
             return ResponseEntity.notFound().build();
         }
 
+        // 3. 유저가 프로젝트에 속해 있는지 확인
+        boolean isUserInProject = projectService.isUserInProject(user, project);
+        if (!isUserInProject) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         // 프로젝트헤 해당하는 할 일 목록 조회
         List<ProjectTodo> todos = projectTodoRepository.findByProjectId(projectId);
         if (todos.isEmpty()) {
@@ -176,13 +182,18 @@ public class ProjectTodoController {
 
             ProjectTodo projectTodo = optionalProjectTodo.get();
 
-            // 4. 프로젝트 할 일에 연관된 멤버 정보 삭제
+            // 5. 사용자가 프로젝트에 속해 있는지 확인
+            if (!projectService.isUserInProject(user, project)) {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
+
+            // 6. 프로젝트 할 일에 연관된 멤버 정보 삭제
             List<ProjectTodoMember> membersToRemove = projectTodoMemberRepository.findByProjectTodo(projectTodo);
             for (ProjectTodoMember member : membersToRemove) {
                 projectTodoMemberRepository.delete(member);
             }
 
-            // 5. 프로젝트 할 일 삭제
+            // 7. 프로젝트 할 일 삭제
             projectTodoRepository.delete(projectTodo);
 
             return ResponseEntity.ok("할 일이 삭제되었습니다.");
